deduplicate convert_type_of_value
