+++ augur/application/config.py
@@ -5,6 +5,7 @@ from typing import List, Any, Optional
 import os
 from augur.application.db.models import Config 
 from augur.application.db.util import execute_session_query
+from augur.application.util import convert_type_of_value
 
 def get_development_flag_from_config():
     
@@ -108,35 +109,6 @@ default_config = {
             }
         }
 
-
-def convert_type_of_value(config_dict, logger=None):
-        
-        data_type = config_dict["type"]
-
-        if data_type == "str" or data_type is None:
-            return config_dict
-
-        elif data_type == "int":
-            config_dict["value"] = int(config_dict["value"])
-
-        elif data_type == "bool":
-            value = config_dict["value"]
-            
-            if value.lower() == "false":
-                config_dict["value"] = False
-            else:
-                config_dict["value"] = True
-
-        elif data_type == "float":
-            config_dict["value"] = float(config_dict["value"])
-
-        else:
-            if logger:
-                logger.error(f"Need to add support for {data_type} types to config") 
-            else:
-                print(f"Need to add support for {data_type} types to config")
-
-        return config_dict
 
 class AugurConfig():
